boom

+++ upgrade-tomcat.sh
@@ -20,15 +20,25 @@
 TOMCAT_VERSION_FILE="/opt/current_tomcat"
 TOMCAT_CURRENT=`cat $TOMCAT_VERSION_FILE`
 TOMCAT_LATEST=`curl -s http://mirrors.ibiblio.org/apache/tomcat/tomcat-9/ |grep v9 | cut -f 3 -d \> | cut -f 1 -d \/`
+SHA_FILE=
 TOMCAT_VERSION=`echo $TOMCAT_LATEST | cut -f 2 -d v`
 TOMCAT_WORKING_DIR="/opt/tomcat9"
 TMP_TOMCAT="/tmp/apache-tomcat-$TOMCAT_VERSION.tar.gz"
+TMP_SHASUM="/tmp/apache-tomcat-$TOMCAT_VERSION.tar.gz.sha512"
 
 if [ "$TOMCAT_CURRENT" -ne "$TOMCAT_LATEST" ];	
 	echo "Our Tomcat: $TOMCAT_CURRENT, Latest Release: $TOMCAT_LATEST"
 	echo "Downloading Tomcat: $TOMCAT_LATEST..."
-	URL = "http://apache.mirrors.hoobly.com/tomcat/tomcat-9/$TOMCAT_LATEST/bin/apache-tomcat-$TOMCAT_VERSION.tar.gz"
-	curl $URL -o $TMP_TOMCAT
+	curl http://apache.mirrors.hoobly.com/tomcat/tomcat-9/$TOMCAT_LATEST/bin/apache-tomcat-$TOMCAT_VERSION.tar.gz -o $TMP_TOMCAT
+	curl https://www.apache.org/dist/tomcat/tomcat-9/$TOMCAT_LATEST/bin/apache-tomcat-$TOMCAT_VERSION.tar.gz.sha512 -o $TMP_SHASUM
+	SHA_SUM = `sha512sum $TMP_TOMCAT`
+	ACTUAL_SUM = `cat $TMP_SHASUM`
+
+	if [ "$SHA_SUM" -ne "$ACTUAL_SUM"];
+		echo "SHA sums do not match. Exiting..."
+		exit
+	fi
+
 	echo "Done!"
 	echo "Stopping CAS..."
 	systemctl stop tomcat9
